+++ src/commands/types.ts
@@ -9,6 +9,8 @@ export type BaseChartOptions = {
   removeFirstTicks: number;
   maxTicks: number;
   trimPrefix: string;
+  customNames: Map<string, string>;
+  namesFile: string;
   aggregateFile: string;
   stddevFilter: number;
   metrics: MetricEnum[];